refactor: components and modules folder
