+++ ui/src/app/guards/google-oauth.guard.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
-import { UserService } from 'src/app/modules/home/services/user.service';
+import { UserService } from 'src/app/components/home/services/user.service';
 
 @Injectable({
   providedIn: 'root'
