+++ ui/src/app/shared/dialogs/confirmation-dialog/confirmation-dialog.component.html
