+++ ui/src/app/shared/dialogs/subject-details-dialog/subject-details-dialog.component.spec.ts
