+++ ui/src/app/shared/export-calendar/components/calendar-options/calendar-options.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, NgZone } from '@angular/core';
 import { Router } from '@angular/router';
-import { UserService } from 'src/app/modules/home/services/user.service';
+import { UserService } from 'src/app/components/home/services/user.service';
 
 @Component({
   selector: 'app-calendar-options',
