+++ ui/src/app/shared/export-calendar/components/subjects-selector/subjects-selector.component.ts
@@ -1,10 +1,10 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subject } from 'src/app/models/subject.model';
 import { FormGroup, FormBuilder, FormArray } from '@angular/forms';
-import { UserService } from 'src/app/modules/home/services/user.service';
-import { GoogleCalendarService } from 'src/app/modules/home/services/google-calendar.service';
+import { UserService } from 'src/app/components/home/services/user.service';
+import { GoogleCalendarService } from 'src/app/components/home/services/google-calendar.service';
 import { MatDialog } from '@angular/material/dialog';
-import { SubjectDetailsDialogComponent } from 'src/app/components/dialogs/subject-details-dialog/subject-details-dialog.component';
+import { SubjectDetailsDialogComponent } from 'src/app/shared/dialogs/subject-details-dialog/subject-details-dialog.component';
 import { NotificationService } from 'src/app/services/notification.service';
 
 @Component({
