+++ ui/src/app/shared/period-selector/period-selector.component.spec.ts
