+++ ui/src/app/shared/schedule/schedule.component.html
